dashboard

+++ database/migrations/2021_09_03_093712_create_ingredient_boxes_table.php
@@ -14,13 +14,10 @@ class CreateIngredientBoxesTable extends Migration
     public function up()
     {
         Schema::create('ingredient_boxes', function (Blueprint $table) {
-            $table->float('quantity', 4, 2);
+            $table->integer('quantity');
             $table->date('expiration_date');
             $table->foreignId('user_id')->constrained();
             $table->foreignId('ingredient_id')->constrained();
-
-            // $table->unsignedBigInteger('ingredient_id');
-            // $table->foreign('ingredient_id')->references('id')->on('ingredients');
             $table->timestamps();
             
         });
